refactored OnMouseEventSample
